Will use localhost later
Found out that it is easier to simulate a game from one window and
change the view (graphics) ever time a player ends its turn. This will
later on be easy to change for the online muliplayer mode, whilst it's
easier to test things when the whole game is being played in just on
window on one computer.

Much code has been added, but it should be easy to understand the
intention behind the creation of the new classes.