Added objects to PlayerPanel
Graphics for objects were made and added to their respective classas.
The objects was also added to the PlayerPanel class.

Next step is to track how many of each object is available for the
current player, as well as implementing method for change of turns, and
thereby player.

Later on, also enable players to move and places cities and roads, and
paint graphics to the board and store nessecary data and check for rule
violations.

Added pieces:
- Road
- City
Added ResorceCards:
-hay
-clay
-sheep
-rock
-wood